website strucutre ok with new routes and links
