+++ src/assets/galactic_vanguardian_ncv.otf
