+++ src/components/Header.css
@@ -9,6 +9,7 @@
 
 #logo {
   margin: 10px 20px 10px 20px;
+  height: 50px;
 }
 
 #logBtn {
