+++ src/components/Header.jsx
@@ -1,13 +1,16 @@
-import logo from '../assets/logoV2.png';
-
-import '../wll.css';
+import { Link } from 'react-router-dom';
+import './Header.css';
 import '../index.css';
+import logo from '../assets/logoV2.png';
 
 const Header = () => {
   return (
-    <div>
+    <header>
       <nav className="navBar">
-        <img id="logo" src={logo} alt="Logo" />
+        <Link to="/">
+          {' '}
+          <img id="logo" src={logo} alt="Logo" />{' '}
+        </Link>
         <button
           className="bg-melon-pink hover:bg-dark-sienna text-black hover:text-white font-bold px-2 py-2 m-5"
           type="button"
@@ -16,7 +19,7 @@ const Header = () => {
           LOGiN
         </button>
       </nav>
-    </div>
+    </header>
   );
 };
