--- src/components/routes/Accueil.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const accueil = () => {
-  return <div>Page d'accueil</div>;
-};
-
-export default accueil;
