+++ src/components/routes/Destination.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Destination = () => {
+  return <div>page destination</div>;
+};
+
+export default Destination;
