+++ src/components/routes/Home.css
@@ -1,29 +1,3 @@
-@font-face {
-  font-family: 'title';
-  src: url('./assets/GALACTIC\ VANGUARDIAN\ NCV.otf');
-}
-:root {
-  --lightestShade: #ffe4d8;
-  --lightShade: #fcb9b2;
-  --accentShade: #b23a48;
-  --drakestShade: #461220;
-  --smallBorderRadius: 0.3rem;
-  background-color: var(--lightestShade);
-  font-size: 16px;
-}
-
-h1 {
-  font-family: 'title', sans-serif;
-  color: var(--drakestShade);
-  font-size: 1.5rem;
-}
-
-h2 {
-  color: var(--accentShade);
-  font-size: 1.2rem;
-  margin-top: -0.5rem;
-}
-
 #country-input {
   width: 100%;
   height: 2.5rem;
