+++ src/components/routes/PostDetails.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const PostDetails = () => {
+  return <div>WIP</div>;
+};
+
+export default PostDetails;
