--- src/components/routes/ResultsPage.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ResultsPage = () => {
-  return <div>Page de résultats</div>;
-};
-
-export default ResultsPage;
