+++ src/index.css
@@ -1,3 +1,46 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+html,
+body,
+#root {
+  height: 100%;
+}
+
+body,
+#root {
+  margin: O;
+  display: flex;
+  flex-direction: column;
+}
+
+main {
+  flex-grow: 1;
+}
+
+@font-face {
+  font-family: 'title';
+  src: url('./assets/galactic_vanguardian_ncv.otf');
+}
+:root {
+  --lightestShade: #ffe4d8;
+  --lightShade: #fcb9b2;
+  --accentShade: #b23a48;
+  --drakestShade: #461220;
+  --smallBorderRadius: 0.3rem;
+  background-color: var(--lightestShade);
+  font-size: 16px;
+}
+
+h1 {
+  font-family: 'title', sans-serif;
+  color: var(--drakestShade);
+  font-size: 1.5rem;
+}
+
+h2 {
+  color: var(--accentShade);
+  font-size: 1.2rem;
+  margin-top: -0.5rem;
+}
